services/nlp-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
spacy==3.7.2
scikit-learn==1.3.2
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
psycopg2-binary==2.9.9
redis==5.0.1
httpx==0.25.2
python-dotenv==1.0.0
